+++ packages/plugin-mermaid/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @infinite-canvas-tutorial/mermaid
 
+## 0.0.7
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @infinite-canvas-tutorial/webcomponents@0.0.8
+    -   @infinite-canvas-tutorial/ecs@0.0.6
+
 ## 0.0.6
 
 ### Patch Changes