+++ packages/webcomponents/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @infinite-canvas-tutorial/webcomponents
 
+## 0.0.8
+
+### Patch Changes
+
+-   Lasso draw mode.
+-   Updated dependencies
+    -   @infinite-canvas-tutorial/ecs@0.0.6
+
 ## 0.0.7
 
 ### Patch Changes